add initial routes for subtypes
